add script to export a dataset in manual import format
Adds the `create_manual_import.py` script, which
converts a folder's output to a single csv for manual
import to https://owid.cloud/admin/import. This script
can be useful for debugging purposes where you want
to upload a dataset to the Grapher, but are not yet
ready to run `import_dataset` on the dataset.