fix(import_dataset): delete existing data values before insert
Fixes a bug where old data values may not be deleted/overwritten when
an importer is run a second time with different data.

e.g. when the importer is first run, there is an observation for Canada
in 2019 for variableId=1. But then some
fixes are made to the importer and in the updated data, there is no
longer a Canada, 2019 observation for this variable. If the importer is
run now, the existing Canada, 2019 observation will not be deleted.